Added mbaas and core tests

+++ utils/routes.js
@@ -1,13 +1,22 @@
 const exec = require('./execute');
 
 async function exposeService(name) {
+  if (name === 'fh-mbaas') {
+    return await exec(`oc get routes | awk '/${name}/{print $2}'`);
+  }
   try {
-    await exec(`oc expose service ${name}`);
-  } catch (_) {}  // eslint-disable-line no-empty
+    const serviceName = await exec(`oc get services | awk '/^${name}/{print $1}'`);
+    await exec(`oc expose service ${serviceName}`);
+  } catch (error) {
+    console.error(error);
+  }
   return await exec(`oc get routes | awk '/^${name}/{print $2}'`);
 }
 
 async function deleteRoute(name) {
+  if (name === 'fh-mbaas') {
+    return;
+  }
   try {
     const routeName = await exec(`oc get routes | awk '/^${name}/{print $1}'`);
     await exec(`oc delete route ${routeName}`);
